update destroy script

+++ destroy_vSAN-Cluster.ps1
@@ -19,10 +19,12 @@ $cluster | Get-VMHost | Set-VMHost -State Disconnected -Confirm:$false
 $cluster | Get-VMHost | Remove-VMHost -Confirm:$false
 $cluster | Remove-Cluster -Confirm:$false
 # Remove Witness Host / DC
-Get-VMHost $vsan_witness_host | Set-VMHost -State Disconnected -Confirm:$false
-Get-VMHost $vsan_witness_host | Remove-VMHost -Confirm:$false
-if((Get-Datacenter $witness_dc | Get-VMHost).Count -eq 0){
-    Remove-Datacenter $witness_dc -Confirm:$false
+if($vsan_witness_host){
+    Get-VMHost $vsan_witness_host | Set-VMHost -State Disconnected -Confirm:$false
+    Get-VMHost $vsan_witness_host | Remove-VMHost -Confirm:$false
+    if((Get-Datacenter $witness_dc | Get-VMHost).Count -eq 0){
+        Remove-Datacenter $witness_dc -Confirm:$false
+    }    
 }
 Disconnect-VIServer * -Confirm:$false
 
@@ -30,6 +32,8 @@ Connect-VIServer -Server $base_vc_address `
     -User $base_vc_user -Password $base_vc_pass -Force
 Get-VM $vm_name_list | Stop-VM -Confirm:$false
 Get-VM $vm_name_list | Remove-VM -DeletePermanently -Confirm:$false
-Get-VM $vsan_witness_va_name | Stop-VM -Confirm:$false
-Get-VM $vsan_witness_va_name | Remove-VM -DeletePermanently -Confirm:$false
+if($vsan_witness_va_name){
+    Get-VM $vsan_witness_va_name | Stop-VM -Confirm:$false
+    Get-VM $vsan_witness_va_name | Remove-VM -DeletePermanently -Confirm:$false    
+}
 Disconnect-VIServer * -Confirm:$false
